batman-adv: GLUON: assume absence of mcast routers on
In Gluon networks we currently have the issue that a lot of devices are
still running OpenWrt < 23.05. And likely many small RAM/flash devices
will be stuck on Gluon v2021.1.2 / OpenWrt 19.07 for a while longer.

These devices would block the availability of the batman-adv multicast
opitimizations feature for routable IPv6 multicast addresses. Therefore,
instead of defaulting to "multicast-router-present" this changes the
behaviour to assume "no-multicast-router-present" on this "old" devices
/ kernels.

The only downside of this is: If there were a multicast router behind
one of these "old" devices / kernels then this would potentially lead to
multicast packet loss to it. However, with the current
"multicast-router-present" default we would likely get over the 16 nodes
limit enforced in Gluon (noflood-mark) and would have such packets
filtered anyway. Therefore there is no real downside for this change in
Gluon.

Also the compat fallback / old mechanism to check for a listener for the
IPv6 link-local multicast all-routers address (ff02::2) to detect
multicast routers won't mitigate this in the Gluon case either, as each
Gluon node is currently configured as a (unicast) router and is therefore
listening on this address. So again we would very likely still get over
the 16 nodes limit.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>