batman-adv: GLUON: pull in gluon-mesh-batman-adv-brmldproxy on OpenWrt 23.05
This feeds repository is supposed to be a drop-in replacement / update for
batman-adv in Gluon. That is for ease of use just adding this repository
to Gluon's site's modules file should be enough to update batman-adv /
Gluon with the latest multicast related features. Without needing to
touch a site.mk.

Later, once all this is merged upstream in Gluon then this dependency
should be part of Gluon's package/features, similar to how
gluon-ebtables-limit-arp is handled and included by default. Instead of
hacking this DEPENDS into the batman-adv package.

Note: the brmldproxy currently won't work on OpenWrt < 23.05. Therefore
only including it on OpenWrt 23.05 / Linux 5.15.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>